# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Satara Development Internal Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feasibility:
    get:
      operationId: listFeasibilityStudies
      tags: [feasibility]
      summary: List feasibility studies
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeasibilityStudy"
    post:
      operationId: createFeasibilityStudy
      tags: [feasibility]
      summary: Create feasibility study
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeasibilityStudyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeasibilityStudy"
  /feasibility/{id}:
    get:
      operationId: getFeasibilityStudy
      tags: [feasibility]
      summary: Get feasibility study
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feasibility study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeasibilityStudy"
    patch:
      operationId: updateFeasibilityStudy
      tags: [feasibility]
      summary: Update feasibility study
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeasibilityStudyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeasibilityStudy"

  # ─── LEGAL ───────────────────────────────────────────────────
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: kprStatus
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/kpr-pipeline:
    get:
      operationId: getKprPipeline
      tags: [customers]
      summary: KPR pipeline summary
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: KPR pipeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KprPipeline"

  # ─── UNITS ───────────────────────────────────────────────────
...
  /construction/tasks:
    get:
      operationId: listConstructionTasks
      tags: [construction]
      summary: List construction tasks
      parameters:
        - name: unitId
          in: query
          schema:
            type: ["integer", "null"]
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConstructionTask"
    post:
      operationId: createConstructionTask
      tags: [construction]
      summary: Create construction task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstructionTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructionTask"
  /construction/tasks/{id}:
    patch:
      operationId: updateConstructionTask
      tags: [construction]
      summary: Update construction task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstructionTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructionTask"
  /construction/progress-summary:
    get:
      operationId: getConstructionProgressSummary
      tags: [construction]
      summary: Construction progress summary by project
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Progress summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressSummary"

  # ─── QC ───────────────────────────────────────────────────────
...
  /handovers:
    get:
      operationId: listHandovers
      tags: [handovers]
      summary: List handovers
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of handovers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Handover"
    post:
      operationId: createHandover
      tags: [handovers]
      summary: Create handover (BAST)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Handover"
  /handovers/{id}:
    get:
      operationId: getHandover
      tags: [handovers]
      summary: Get handover detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Handover
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Handover"
components:
  schemas:
...
    FeasibilityStudy:
      type: object
      required: [id, projectId, hpp, roi, margin, isApproved, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        hpp: { type: number }
        roi: { type: number }
        margin: { type: number }
        cashflow: { type: ["number", "null"] }
        bep: { type: ["number", "null"] }
        rab: { type: ["number", "null"] }
        rabVariance: { type: ["number", "null"] }
        isApproved: { type: boolean }
        approvedAt: { type: ["string", "null"] }
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
    FeasibilityStudyInput:
      type: object
      required: [projectId, hpp, roi, margin]
      properties:
        projectId: { type: integer }
        hpp: { type: number }
        roi: { type: number }
        margin: { type: number }
        cashflow: { type: number }
        bep: { type: number }
        rab: { type: number }
        catatan: { type: string }
    FeasibilityStudyUpdate:
      type: object
      properties:
        hpp: { type: number }
        roi: { type: number }
        margin: { type: number }
        cashflow: { type: number }
        bep: { type: number }
        rab: { type: number }
        rabVariance: { type: number }
        isApproved: { type: boolean }
        catatan: { type: string }

    # ─── LEGAL ───────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, nama, nik, kontak, statusKpr, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: ["integer", "null"] }
        unitId: { type: ["integer", "null"] }
        nama: { type: string }
        nik: { type: string }
        kontak: { type: string }
        pekerjaan: { type: ["string", "null"] }
        bank: { type: ["string", "null"] }
        statusKpr:
          type: string
          enum: [bi_checking, collect_berkas, input_bank, follow_up, sp3k, akad, selesai, ditolak]
        berkasLengkap: { type: boolean }
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
    CustomerInput:
      type: object
      required: [nama, nik, kontak, statusKpr]
      properties:
        projectId: { type: integer }
        unitId: { type: integer }
        nama: { type: string }
        nik: { type: string }
        kontak: { type: string }
        pekerjaan: { type: string }
        bank: { type: string }
        statusKpr:
          type: string
          enum: [bi_checking, collect_berkas, input_bank, follow_up, sp3k, akad, selesai, ditolak]
        catatan: { type: string }
    CustomerUpdate:
      type: object
      properties:
        unitId: { type: integer }
        bank: { type: string }
        statusKpr:
          type: string
          enum: [bi_checking, collect_berkas, input_bank, follow_up, sp3k, akad, selesai, ditolak]
        berkasLengkap: { type: boolean }
        catatan: { type: string }
    KprPipeline:
      type: object
      required: [total, byStatus]
      properties:
        total: { type: integer }
        byStatus:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status: { type: string }
              count: { type: integer }
        rejectRate: { type: number }
        avgDuration: { type: number }

    # ─── UNIT ────────────────────────────────────────────────────
...
    ConstructionTask:
      type: object
      required: [id, unitId, item, bobot, status, createdAt]
      properties:
        id: { type: integer }
        unitId: { type: integer }
        item: { type: string }
        bobot: { type: number }
        status:
          type: string
          enum: [belum_mulai, dalam_proses, selesai]
        tanggalMulai: { type: ["string", "null"] }
        tanggalSelesai: { type: ["string", "null"] }
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
    ConstructionTaskInput:
      type: object
      required: [unitId, item, bobot]
      properties:
        unitId: { type: integer }
        item: { type: string }
        bobot: { type: number }
        status:
          type: string
          enum: [belum_mulai, dalam_proses, selesai]
        tanggalMulai: { type: string }
    ConstructionTaskUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [belum_mulai, dalam_proses, selesai]
        tanggalMulai: { type: string }
        tanggalSelesai: { type: string }
        catatan: { type: string }
    ProgressSummary:
      type: object
      required: [unitId, unitLabel, progress, deviation, readyAkad]
      properties:
        unitId: { type: integer }
        unitLabel: { type: string }
        progress: { type: number }
        deviation: { type: number }
        readyAkad: { type: boolean }
        openDefects: { type: integer }

    # ─── QC ───────────────────────────────────────────────────────
...
    Handover:
      type: object
      required: [id, unitId, customerId, tanggal, skorKepuasan, createdAt]
      properties:
        id: { type: integer }
        unitId: { type: integer }
        customerId: { type: integer }
        tanggal: { type: string }
        skorKepuasan: { type: ["number", "null"] }
        bastGenerated: { type: boolean }
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
    HandoverInput:
      type: object
      required: [unitId, customerId, tanggal]
      properties:
        unitId: { type: integer }
        customerId: { type: integer }
        tanggal: { type: string }
        skorKepuasan: { type: number }
        catatan: { type: string }