# lib/api-zod/src/generated/types/feasibilityStudyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Satara Development Internal Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface FeasibilityStudyInput {
  projectId: number;
  hpp: number;
  roi: number;
  margin: number;
  cashflow?: number;
  bep?: number;
  rab?: number;
  catatan?: string;
}